requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.2.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
av==9.1.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.2
bitsandbytes==0.41.3
bleach==4.1.0
blinker==1.6.2
bracex==2.4
build==1.2.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.10
chroma-hnswlib==0.7.3
chromadb==0.5.0
ci-info==0.3.0
click==8.1.6
coloredlogs==15.0.1
comm==0.2.0
configobj==5.0.8
configparser==6.0.1
customtkinter==5.2.1
cycler==0.11.0
darkdetect==0.8.0
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
Django==4.2.5
django-model-utils==4.3.1
entrypoints==0.3
etelemetry==0.3.1
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.12.2
fitz==0.0.1.dev2
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.28.3
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
geocoder==1.38.1
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.6.4
google-api-core==2.12.0
google-api-python-client==2.133.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gradio==4.25.0
gradio_client==0.15.0
greenlet==2.0.2
groq==0.9.0
grpcio==1.59.0
grpcio-status==1.58.0
grpcio-tools==1.59.0
gTTS==2.5.1
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.3
importlib-metadata==6.8.0
importlib_resources==6.4.0
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.1.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kafka-python==2.0.2
keras==2.13.1
kiwisolver==1.3.2
kubernetes==29.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.0
langsmith==0.1.67
libclang==16.0.6
llvmlite==0.42.0
looseversion==1.3.0
lxml==5.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.5.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mistune==0.8.4
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
MYLIB @ file:///Users/ayonsomaddar/Chandigarh%20University/Ayon/Mylib
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numba==0.59.1
numexpr==2.8.5
numpy==1.22.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt-einsum==3.3.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pbr==6.0.0
pcapng==0.1.25
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
poodle==1.3.3
posthog==3.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf==4.23.4
prov==2.0.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.15
pydantic_core==2.16.3
pydeck==0.8.0
pydot==2.0.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.0.6
pypdf==3.15.4
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pyrsistent==0.18.0
pyshark==0.6
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.9
python-pcapng==2.1.1
python-utils==3.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.1
qiskit==1.0.1
QtPy==1.11.3
ratelim==0.1.6
rdflib==7.0.0
regex==2023.8.8
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
ruff==0.3.5
rustworkx==0.14.1
safetensors==0.4.2
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-scm==6.3.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
stevedore==5.2.0
streamlit==1.25.0
streamlit-space==0.1.5
streamlit-toggle==0.1.3
streamlit-toggle-switch==1.0.2
symengine==0.11.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.12.1
testpath==0.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.0.0
tiktoken==0.4.0
tkintermapview==1.29
tkvideoplayer==2.5
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
traits==6.3.2
transformers==4.37.2
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==4.3.1
ucimlrepo==0.0.3
ultralytics==8.0.226
uritemplate==4.1.1
urllib3==1.26.8
utils==1.0.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.21.2
watchfiles==0.22.0
wcmatch==8.5.1
wcwidth==0.2.10
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==3.5.2
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
